Applied modifications for map bottom sheet's view (abbreviations)

+++ app/src/main/java/com/emplk/realestatemanager/data/property/PropertyTypeSurfacePriceAndPictureDto.kt
@@ -7,7 +7,7 @@ data class PropertyTypeSurfacePriceAndPictureDto(
     val id: Long,
     val type: String,
     val price: BigDecimal,
-    val surface: Int,
+    val surface: Double,
     @ColumnInfo(name = "uri")
     val pictureUri: String,
     /*   @ColumnInfo(name = "name")
